Add header component; Make code refactoring
