+++ src/components/pages/favorite-places-page/fevorite-places-page.jsx
@@ -1,31 +1,12 @@
 import React from 'react';
-import {Link} from "react-router-dom";
+import Header from "../../layout/header/header";
+import {MY_ONLY_USER} from "../../../mock/users";
 
 const FavoritePlacesPage = () => {
   return (
     <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to="/">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <a className="header__nav-link header__nav-link--profile" href="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                  </a>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+
+      <Header userName={MY_ONLY_USER}/>
 
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
