+++ src/mock/users.js
@@ -1,0 +1,1 @@
+export const MY_ONLY_USER = `Oliver.conner@gmail.com`;
